# docs/openapi/andy-policies-v1.yaml
openapi: 3.0.1
info:
  title: Andy Policies API
  description: 'Governance policy catalog — structured, versioned policy documents with lifecycle and audit trail, consumed by Conductor for story admission, verification, and compliance reporting (content only; enforcement lives in consumers).'
  contact:
    name: Rivoli AI
    url: https://github.com/rivoli-ai/andy-policies
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: v1
paths:
...
  /api/bundles:
    post:
      tags:
        - Bundles
      summary: "Create a new bundle (frozen snapshot of the live catalog).\r\nReturns 201 with the Andy.Policies.Application.Interfaces.BundleDto + `Location`\r\nheader pointing at `GET /api/bundles/{id}`.\r\nMaps Andy.Policies.Application.Exceptions.ValidationException → 400 and\r\nAndy.Policies.Application.Exceptions.ConflictException → 409 (duplicate active name).\r\nP8.6 (#86) — the CLI from this story is the first REST consumer."
      operationId: Bundles_Create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBundleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateBundleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateBundleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      tags:
        - Bundles
      summary: "List bundles. `?includeDeleted=true` includes soft-deleted\r\nrows; `take` is clamped server-side. P8.6 (#86)."
      operationId: Bundles_List
      parameters:
        - name: includeDeleted
          in: query
          schema:
            type: boolean
            default: false
        - name: skip
          in: query
          schema:
            type: integer
            format: int32
            default: 0
        - name: take
          in: query
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BundleDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/bundles/{id}':
    get:
      tags:
        - Bundles
      summary: "Get a bundle by id. Returns 200 with Andy.Policies.Application.Interfaces.BundleDto or\r\n404 when the bundle is missing. Soft-deleted bundles are\r\naddressable here (the row remains for audit-chain integrity);\r\npass `?includeDeleted=true` on `GET /api/bundles` to\r\nlist them. P8.6 (#86)."
      operationId: Bundles_Get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Bundles
      summary: "Soft-delete a bundle. State flips to\r\nAndy.Policies.Domain.Enums.BundleState.Deleted; the row remains in the table\r\nfor audit-chain integrity. Idempotent: a second delete on the\r\nsame id returns 404. P8.6 (#86)."
      operationId: Bundles_Delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rationale
          in: query
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/bundles/{id}/resolve':
    get:
      tags:
        - Bundles
      summary: "Resolve bindings for a `(targetType, targetRef)` pair\r\nagainst a frozen bundle. Exact-match only (mirrors the live\r\n`GET /api/bindings/resolve`); no hierarchy walk.\r\nReturns 200 with a Andy.Policies.Application.Interfaces.BundleResolveResult; 404\r\nwhen the bundle does not exist or is soft-deleted; 400 when\r\ntargetRef is empty."
      operationId: Bundles_Resolve
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: targetType
          in: query
          schema:
            $ref: '#/components/schemas/BindingTargetType'
        - name: targetRef
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            ETag:
              description: 'Strong validator: "<snapshotHash>" (64 hex chars). Bundles are immutable post-insert (P8.1), so the ETag is stable for the lifetime of the bundle id.'
              schema:
                type: string
            Cache-Control:
              description: 'public, max-age=31536000, immutable — the response body for a given (bundleId, snapshotHash) never changes.'
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleResolveResult'
        '304':
          description: Not Modified
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/bundles/{id}/policies/{policyId}':
    get:
      tags:
        - Bundles
      summary: "Look up a single pinned policy by id. Returns 200 with a\r\nAndy.Policies.Application.Interfaces.BundlePinnedPolicyDto, or 404 when the bundle\r\ndoes not exist / is soft-deleted, or the policy id is not in\r\nthe snapshot."
      operationId: Bundles_GetPinnedPolicy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: policyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          headers:
            ETag:
              description: 'Strong validator: "<snapshotHash>" (64 hex chars). Bundles are immutable post-insert (P8.1), so the ETag is stable for the lifetime of the bundle id.'
              schema:
                type: string
            Cache-Control:
              description: 'public, max-age=31536000, immutable — the response body for a given (bundleId, snapshotHash) never changes.'
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundlePinnedPolicyDto'
        '304':
          description: Not Modified
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/bundles/{id}/contents':
    get:
      tags:
        - Bundles
      summary: "P9 follow-up #204 (2026-05-07) — bulk contents tree for the\r\nfrozen-tree view in the bundle detail page. Returns every\r\npolicy in the snapshot with its bindings nested underneath,\r\nplus a flat list of overrides. Same ETag / Cache-Control\r\nposture as the other snapshot-backed reads (immutable, max-age\r\n= 1y) since the response body is bound to a specific snapshot\r\nhash. Returns 404 when the bundle is missing / soft-deleted."
      operationId: Bundles_Contents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleContentsDto'
        '304':
          description: Not Modified
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/bundles/{id}/diff':
    get:
      tags:
        - Bundles
      summary: "Emit an RFC-6902 JSON Patch between two bundles' frozen\r\nsnapshots (P8.6, story rivoli-ai/andy-policies#86). Returns\r\n200 with a Andy.Policies.Application.Interfaces.BundleDiffResult; 404 when either\r\nbundle is missing or soft-deleted; 400 when the same id is\r\npassed for both id and `to` (the\r\ntrivial empty-patch case is allowed via two distinct ids\r\nthat happen to resolve to identical canonical bytes — the\r\ndiff returns `[]`; same-id is rejected as a likely\r\ncaller mistake)."
      operationId: Bundles_Diff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: to
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleDiffResult'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    BindingTargetType:
      enum:
        - Template
        - Repo
        - ScopeNode
        - Tenant
        - Org
        - Agent
      type: string
...
    BundleContentsDto:
      type: object
      properties:
        bundleId:
          type: string
          format: uuid
        bundleName:
          type: string
          nullable: true
        snapshotHash:
          type: string
          nullable: true
        capturedAt:
          type: string
          format: date-time
        policies:
          type: array
          items:
            $ref: '#/components/schemas/BundleContentsPolicyDto'
          nullable: true
        overrides:
          type: array
          items:
            $ref: '#/components/schemas/BundleContentsOverrideDto'
          nullable: true
      additionalProperties: false
      description: "Denormalised contents view of a frozen bundle (P9 follow-up #204,\r\n2026-05-07). Consumers (the bundle detail page) render this as a\r\ntree: policies grouped by name, each expandable to its bindings,\r\nplus a flat list of overrides. `SnapshotHash` is the same\r\nstrong validator the resolution endpoints emit so callers can\r\nshare an HTTP cache between this endpoint and `/resolve`."
...
    BundleContentsPolicyDto:
      type: object
      properties:
        policyId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        policyVersionId:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
        enforcement:
          enum:
            - MUST
            - SHOULD
            - MAY
          type: string
          nullable: true
        severity:
          enum:
            - info
            - moderate
            - critical
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
          nullable: true
        summary:
          type: string
          nullable: true
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/BundleContentsBindingDto'
          nullable: true
      additionalProperties: false
    BundleDiffResult:
      type: object
      properties:
        fromId:
          type: string
          format: uuid
        fromSnapshotHash:
          type: string
          nullable: true
        toId:
          type: string
          format: uuid
        toSnapshotHash:
          type: string
          nullable: true
        rfc6902PatchJson:
          type: string
          nullable: true
        opCount:
          type: integer
          format: int32
      additionalProperties: false
      description: 'Result envelope for M:Andy.Policies.Application.Interfaces.IBundleDiffService.DiffAsync(System.Guid,System.Guid,System.Threading.CancellationToken).'
    BundleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBySubjectId:
          type: string
          nullable: true
        snapshotHash:
          type: string
          nullable: true
        state:
          enum:
            - Draft
            - Active
            - WindingDown
            - Retired
          type: string
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        deletedBySubjectId:
          type: string
          nullable: true
      additionalProperties: false
      description: "Wire shape returned by Andy.Policies.Application.Interfaces.IBundleService. Excludes the\r\n(potentially large) `SnapshotJson` payload — surfaces fetch\r\nthat separately on demand via `BundlesController` in P8.3."
    BundlePinnedPolicyDto:
      type: object
      properties:
        bundleId:
          type: string
          format: uuid
        bundleName:
          type: string
          nullable: true
        snapshotHash:
          type: string
          nullable: true
        capturedAt:
          type: string
          format: date-time
        policyId:
          type: string
          format: uuid
        policyName:
          type: string
          nullable: true
        policyVersionId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          format: int32
        enforcement:
          enum:
            - MUST
            - SHOULD
            - MAY
          type: string
          nullable: true
        severity:
          enum:
            - info
            - moderate
            - critical
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
          nullable: true
        rulesJson:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
      additionalProperties: false
      description: "Single-policy projection from a bundle snapshot. Returned by\r\nM:Andy.Policies.Application.Interfaces.IBundleResolver.GetPinnedPolicyAsync(System.Guid,System.Guid,System.Threading.CancellationToken); carries the\r\nsnapshot coordinates so callers can stamp ETags / cache keys."
    BundleResolveResult:
      type: object
      properties:
        bundleId:
          type: string
          format: uuid
        bundleName:
          type: string
          nullable: true
        snapshotHash:
          type: string
          nullable: true
        capturedAt:
          type: string
          format: date-time
        targetType:
          $ref: '#/components/schemas/BindingTargetType'
        targetRef:
          type: string
          nullable: true
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/BundleResolvedBindingDto'
          nullable: true
        count:
          type: integer
          format: int32
      additionalProperties: false
      description: "Envelope for M:Andy.Policies.Application.Interfaces.IBundleResolver.ResolveAsync(System.Guid,Andy.Policies.Domain.Enums.BindingTargetType,System.String,System.Threading.CancellationToken). Carries\r\nthe bundle identity + snapshot coordinate (Andy.Policies.Application.Interfaces.BundleResolveResult.SnapshotHash,\r\nAndy.Policies.Application.Interfaces.BundleResolveResult.CapturedAt) so callers caching the response have\r\neverything they need to validate it against a re-fetched bundle."
...
    CreateBundleRequest:
      type: object
      properties:
        name:
          type: string
          description: "Slug shape: `^[a-z0-9][a-z0-9-]{0,62}$`.\r\n            Active bundles are unique by name; soft-deleted bundles release\r\n            the slug for reuse."
          nullable: true
        description:
          type: string
          description: Optional human-readable summary.
          nullable: true
        rationale:
          type: string
          description: "Required non-empty rationale recorded\r\n            against the audit event."
          nullable: true
        includeOverrides:
          type: boolean
          description: "When true (P9 follow-up #205,\r\n            2026-05-07), the snapshot also captures the current set of\r\n            Approved, non-expired overrides. Default `true` preserves the\r\n            pre-#205 behaviour where overrides were always included; pass\r\n            `false` for compliance / immutability bundles where overrides\r\n            are intentionally excluded."
      additionalProperties: false
      description: 'Inputs to M:Andy.Policies.Application.Interfaces.IBundleService.CreateAsync(Andy.Policies.Application.Interfaces.CreateBundleRequest,System.String,System.Threading.CancellationToken).'
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]
...